@@ -17,18 +17,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.3.0",
-    "@angular/common": "^14.3.0",
-    "@angular/compiler": "^14.3.0",
-    "@angular/core": "^14.3.0",
+    "@angular/animations": "^15.2.10",
+    "@angular/common": "^15.2.10",
+    "@angular/compiler": "^15.2.10",
+    "@angular/core": "^15.2.10",
     "@angular/fire": "../../dist/packages-dist",
-    "@angular/forms": "^14.3.0",
-    "@angular/platform-browser": "^14.3.0",
-    "@angular/platform-browser-dynamic": "^14.3.0",
-    "@angular/platform-server": "^14.3.0",
-    "@angular/router": "^14.3.0",
-    "@angular/service-worker": "^14.3.0",
-    "@nguniversal/express-engine": "^14.2.3",
+    "@angular/forms": "^15.2.10",
+    "@angular/platform-browser": "^15.2.10",
+    "@angular/platform-browser-dynamic": "^15.2.10",
+    "@angular/platform-server": "^15.2.10",
+    "@angular/router": "^15.2.10",
+    "@angular/service-worker": "^15.2.10",
+    "@nguniversal/express-engine": "^15.2.1",
     "cross-fetch": "^3.1.4",
     "express": "^4.15.2",
     "firebase-admin": "^10.0.2",
@@ -39,11 +39,11 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1402.13",
-    "@angular-devkit/build-angular": "^14.2.13",
-    "@angular/cli": "^14.2.13",
-    "@angular/compiler-cli": "^14.3.0",
-    "@nguniversal/builders": "^14.2.3",
+    "@angular-devkit/architect": "^0.1502.11",
+    "@angular-devkit/build-angular": "^15.2.11",
+    "@angular/cli": "^15.2.11",
+    "@angular/compiler-cli": "^15.2.10",
+    "@nguniversal/builders": "^15.2.1",
     "@types/express": "^4.17.0",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
@@ -58,7 +58,7 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^14.2.2",
+    "ng-packagr": "^15.2.2",
     "ts-mixer": "^6.0.0",
     "ts-node": "~9.1.1",
     "tslint": "~6.1.3",
