@@ -1,25 +1,25 @@
 import { InjectionToken, NgModule, Optional } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { provideFirebaseApp, initializeApp } from '@angular/fire/app';
+import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
+import { CustomProvider, ReCaptchaV3Provider, initializeAppCheck, provideAppCheck } from '@angular/fire/app-check';
 import { FunctionsModule } from '@angular/fire/functions';
-import { initializeAppCheck, provideAppCheck, ReCaptchaV3Provider, CustomProvider } from '@angular/fire/app-check';
+import { BrowserModule } from '@angular/platform-browser';
 
+import { environment } from '../environments/environment';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { environment } from '../environments/environment';
-import { HomeComponent } from './home/home.component';
-import { UpboatsComponent } from './upboats/upboats.component';
 import { AuthComponent } from './auth/auth.component';
-import { FirestoreComponent } from './firestore/firestore.component';
 import { DatabaseComponent } from './database/database.component';
+import { FirestoreComponent } from './firestore/firestore.component';
 import { FunctionsComponent } from './functions/functions.component';
+import { HomeComponent } from './home/home.component';
 import { MessagingComponent } from './messaging/messaging.component';
 import { RemoteConfigComponent } from './remote-config/remote-config.component';
 import { StorageComponent } from './storage/storage.component';
+import { UpboatsComponent } from './upboats/upboats.component';
 
+import { ScreenTrackingService, UserTrackingService } from '@angular/fire/analytics';
 import type { app } from 'firebase-admin';
 import { AppCheckComponent } from './app-check/app-check.component';
-import { ScreenTrackingService, UserTrackingService } from '@angular/fire/analytics';
 
 export const FIREBASE_ADMIN = new InjectionToken<app.App>('firebase-admin');
 
@@ -41,6 +41,8 @@ export const FIREBASE_ADMIN = new InjectionToken<app.App>('firebase-admin');
     BrowserModule.withServerTransition({ appId: 'serverApp' }),
     AppRoutingModule,
     FunctionsModule,
+  ],
+  providers: [
     provideFirebaseApp(() => initializeApp(environment.firebase)),
     provideAppCheck((injector) =>  {
       const admin = injector.get<app.App|null>(FIREBASE_ADMIN, null);
@@ -57,8 +59,6 @@ export const FIREBASE_ADMIN = new InjectionToken<app.App>('firebase-admin');
         return initializeAppCheck(undefined, { provider, isTokenAutoRefreshEnabled: true });
       }
     }, [new Optional(), FIREBASE_ADMIN]),
-  ],
-  providers: [
     ScreenTrackingService,
     UserTrackingService,
   ],
